FCT: retrofitListe

+++ code-4n5/RetrofitListe/app/src/main/java/org/ye/Personne.kt
@@ -3,5 +3,5 @@ package org.ye
 class Personne {
     var a: Int? = null
     var b: String? = null
-    var c: Array<Int>? = null
+    var c: Array<Long>? = null
 }
